# Optimise memory allocations in `HtmlRenderer`.